backend/src/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
zipp==3.9.0
whitenoise==5.3.0